Fix issue with recursive git discovery
There is a package in the ones installed by npm that
recursively looks for git references in it's install
and fails to install if the repository is a submodule
annd the parent repository cannot be accessed (which
is the case inside the docker container where the
parent repository is not available). The
issue is prevented by removing all hidden git files
so when this package recursive git-based discovery runs
there is nothing to look for.